GameView.c
gameView function added. It should be pretty much done but it does require an updateTraps and updateVampire function (see comments  on lines 274-290 for information on how they should be ideally set out.
Also see line 143: do we need to include when a hunter enters a city in draculas trail? is this needed or would this be in hunters view?